requirements.txt
fastapi
uvicorn[standard]
pillow
requests
tensorflow
numpy
transformers
torch
safetensors
fastapi
